Move things around and get rid of GraphQL Client
