+++ crates/sui-graphql-client/queries/epoch_total_checkpoints.graphql
@@ -1,0 +1,5 @@
+query EpochTotalCheckpoints($id: UInt53!){
+  epoch(id: $id) {
+    totalCheckpoints
+  }
+}
